backend/requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
openai>=1.12.0,<2.0.0
python-multipart>=0.0.9
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0
